Move password token gen to userController
