+++ api/v1/services/index.js
@@ -2,5 +2,6 @@ module.exports = {
 	AuthService: require('./AuthService'),
 	MailService: require('./MailService'),
 	PermissionService: require('./PermissionService'),
-	UserService: require('./UserService')
+	UserService: require('./UserService'),
+	TokenService: require('./TokenService')
 };
